Plan 9 from Bell Labs Fourth Edition 2007-02-23

+++ sys/src/cmd/cfs/disk.h
@@ -6,8 +6,8 @@ typedef struct Disk	Disk;
 struct Disk
 {
 	Bcache;
-	int	nb;	/* number of blocks */
-	int	nab;	/* number of allocation blocks */
+	ulong	nb;	/* number of blocks */
+	ulong	nab;	/* number of allocation blocks */
 	int	b2b;	/* allocation bits to a block */
 	int	p2b;	/* Dptr's per page */
 	char	name[KNAMELEN];
